fix(backend,sf-migrate): SF Number(scale=0) 가정 정정 — 22컬럼 NUMERIC + entity BigDecimal 통일
배경:
  SF describe 의 scale=0 메타가 강제력 없음 (Apex/Data Loader/API INSERT 로 소수 저장 가능).
  실측: Claim.DKRetail__Quantity__c (SF scale=0) 에 0.3 존재 → 기존 BIGINT 컬럼 cast 실패.
  V137 (ErpOrder.order_sales_amount), V138 (ErpOrderProduct 6컬럼) 의 sf-meta-diff 작업이
  "SF scale=0 = 정수 도메인" 가정으로 BIGINT 변환했으나 운영 데이터와 안 맞음 — 본 마이그레이션이
  그 가정을 정정.

backend:
  - V168 — 22컬럼 INTEGER/BIGINT → NUMERIC (7 테이블).
    account(1) / claim(2) / erp_order(1) / erp_order_product(6) /
    monthly_female_employee_integration_schedule(3) / order_request_product(2) /
    promotion_employee(7).
  - Entity 7개 의 var 타입 Long/Int → BigDecimal.
  - 영향 호출처 정정 (DTO 11개 / Service 5개) — 산술 비교 / DTO 직렬화 / SAP payload 등.
  - PromotionEmployee.dkDailyActualSalesAmount: Long → BigDecimal (Formula 재현 유지).
  - ErpOrderUpsertService.parseAmountLong: Long → BigDecimal 반환 (반올림 제거 — NUMERIC 보존).

sf-migrate:
  - FieldMapping 에서 isIntegerScale = false 메타 제거 (값 기반 cast 분기 폐기).
  - migrate-stage1.main.kts streaming insert 의 cast 분기 제거 — 모든 numeric 컬럼이
    NUMERIC 으로 통일되어 PostgreSQL `stringtype=unspecified` 가 자동 cast.
  - 주석 v9 → v10.

test (25 파일):
  - fixture / helper / assertion 정정 — Long literal → BigDecimal.valueOf(NL),
    isEqualTo(NL) → isEqualByComparingTo(BigDecimal.valueOf(NL)).
  - UserEntityTest: employee_code nullable 단언 갱신 (V167 정정).
  - PushMessageReceiverSFAnnotationTest: mapping size 6 → 8 (CreatedDate/LastModifiedDate 추가 반영).

검증:
  - ./gradlew compileKotlin 통과
  - ./gradlew compileTestKotlin 통과
  - ./gradlew test 통과 (3045 tests)

+++ backend/src/main/kotlin/com/otoki/powersales/account/dto/request/AdminAccountUpdateRequest.kt
@@ -131,7 +131,7 @@ data class AdminAccountUpdateRequest(
 
     val annualRevenue: BigDecimal? = null,
 
-    val numberOfEmployees: Int? = null,
+    val numberOfEmployees: BigDecimal? = null,
 
     @field:Size(max = 20, message = "담당 영업사원 사번은 20자 이하여야 합니다.")
     val employeeCode: String? = null,
